# apps/worker/openapi-restful.yaml
openapi: 3.0.3
info:
  title: Sutradhar RESTful Actions API
  description: |
    RESTful API endpoints for GitHub, Slack, and Google Calendar operations.
    
    All endpoints follow REST principles with resource-based URLs, proper HTTP methods,
    and consistent response formats.
  version: 1.0.0
servers:
  - url: http://localhost:2198/api/v1
    description: Development server
...
paths:
...
  /github/repos/{repoSlug}/issues:
    get:
      tags: [GitHub]
      summary: List issues
      parameters:
        - name: repoSlug
          in: path
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
            enum: [open, closed, all]
            default: open
        - name: label
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      issues:
                        type: array
                        items:
                          type: object
                      count:
                        type: integer

    post:
      tags: [GitHub]
      summary: Create issue
      parameters:
        - name: repoSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                body:
                  type: string
                labels:
                  type: array
                  items:
                    type: string
                assignees:
                  type: array
                  items:
                    type: string
                sessionId:
                  type: string
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
...